# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MTAA Fitness API",
        "description": "API for user registration, login, workouts, friends and plans.",
        "contact": {
            "email": "support@fitness-app.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
        "/api/friends/send": {
            "post": {
                "tags": [
                    "Friends"
                ],
                "summary": "Send a friend request",
                "operationId": "547cccb178e04a35d1a076dd2a0ffb14",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Friend request sent"
                    },
                    "400": {
                        "description": "Invalid or duplicate request"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/friends/accept/{id}": {
            "post": {
                "tags": [
                    "Friends"
                ],
                "summary": "Accept a friend request",
                "operationId": "ac3106c9f607080db984184d1da9778b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user who sent the request",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Friend request accepted"
                    },
                    "404": {
                        "description": "Request not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/friends/decline/{id}": {
            "post": {
                "tags": [
                    "Friends"
                ],
                "summary": "Decline a friend request",
                "operationId": "1613224f692fe2c8cba1d475edcbfa4e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user who sent the request",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Friend request declined"
                    },
                    "404": {
                        "description": "Request not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/friends": {
            "get": {
                "tags": [
                    "Friends"
                ],
                "summary": "List friends and friend requests",
                "operationId": "6aa592eb2ff8dbc86578ea9e57a6ab24",
                "responses": {
                    "200": {
                        "description": "List of friends and requests"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/friends/remove/{id}": {
            "delete": {
                "tags": [
                    "Friends"
                ],
                "summary": "Remove a friend",
                "operationId": "aec7386cf969136c23ffd9ba34832f3d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the friend to remove",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Friend removed"
                    },
                    "404": {
                        "description": "Friend not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}